+++ google/cloud/aiplatform_v1beta1/types/endpoint.py
@@ -128,9 +128,9 @@ class DeployedModel(proto.Message):
         id (str):
             Output only. The ID of the DeployedModel.
         model (str):
-            Required. The name of the Model this is the
-            deployment of. Note that the Model may be in a
-            different location than the DeployedModel's
+            Required. The name of the Model that this is
+            the deployment of. Note that the Model may be in
+            a different location than the DeployedModel's
             Endpoint.
         display_name (str):
             The display name of the DeployedModel. If not provided upon
@@ -151,10 +151,11 @@ class DeployedModel(proto.Message):
             ``explanation_spec``
             is not populated, the value of the same field of
             ``Model.explanation_spec``
-            is inherited. The corresponding
+            is inherited. If the corresponding
             ``Model.explanation_spec``
-            must be populated, otherwise explanation for this Model is
-            not allowed.
+            is not populated, all fields of the
+            ``explanation_spec``
+            will be used for the explanation configuration.
         service_account (str):
             The service account that the DeployedModel's container runs
             as. Specify the email address of the service account. If