+++ tests/unit/gapic/aiplatform_v1beta1/test_job_service.py
@@ -117,20 +117,26 @@ def test__get_default_mtls_endpoint():
     assert JobServiceClient._get_default_mtls_endpoint(non_googleapi) == non_googleapi
 
 
-def test_job_service_client_from_service_account_info():
+@pytest.mark.parametrize(
+    "client_class", [JobServiceClient, JobServiceAsyncClient,],
+)
+def test_job_service_client_from_service_account_info(client_class):
     creds = credentials.AnonymousCredentials()
     with mock.patch.object(
         service_account.Credentials, "from_service_account_info"
     ) as factory:
         factory.return_value = creds
         info = {"valid": True}
-        client = JobServiceClient.from_service_account_info(info)
+        client = client_class.from_service_account_info(info)
         assert client.transport._credentials == creds
+        assert isinstance(client, client_class)
 
         assert client.transport._host == "aiplatform.googleapis.com:443"
 
 
-@pytest.mark.parametrize("client_class", [JobServiceClient, JobServiceAsyncClient,])
+@pytest.mark.parametrize(
+    "client_class", [JobServiceClient, JobServiceAsyncClient,],
+)
 def test_job_service_client_from_service_account_file(client_class):
     creds = credentials.AnonymousCredentials()
     with mock.patch.object(
@@ -139,9 +145,11 @@ def test_job_service_client_from_service_account_file(client_class):
         factory.return_value = creds
         client = client_class.from_service_account_file("dummy/file/path.json")
         assert client.transport._credentials == creds
+        assert isinstance(client, client_class)
 
         client = client_class.from_service_account_json("dummy/file/path.json")
         assert client.transport._credentials == creds
+        assert isinstance(client, client_class)
 
         assert client.transport._host == "aiplatform.googleapis.com:443"
 
@@ -506,6 +514,24 @@ def test_create_custom_job_from_dict():
     test_create_custom_job(request_type=dict)
 
 
+def test_create_custom_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.create_custom_job), "__call__"
+    ) as call:
+        client.create_custom_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.CreateCustomJobRequest()
+
+
 @pytest.mark.asyncio
 async def test_create_custom_job_async(
     transport: str = "grpc_asyncio", request_type=job_service.CreateCustomJobRequest
@@ -737,6 +763,22 @@ def test_get_custom_job_from_dict():
     test_get_custom_job(request_type=dict)
 
 
+def test_get_custom_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(type(client.transport.get_custom_job), "__call__") as call:
+        client.get_custom_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.GetCustomJobRequest()
+
+
 @pytest.mark.asyncio
 async def test_get_custom_job_async(
     transport: str = "grpc_asyncio", request_type=job_service.GetCustomJobRequest
@@ -936,6 +978,22 @@ def test_list_custom_jobs(
 
 def test_list_custom_jobs_from_dict():
     test_list_custom_jobs(request_type=dict)
+
+
+def test_list_custom_jobs_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(type(client.transport.list_custom_jobs), "__call__") as call:
+        client.list_custom_jobs()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.ListCustomJobsRequest()
 
 
 @pytest.mark.asyncio
@@ -1266,6 +1324,24 @@ def test_delete_custom_job_from_dict():
     test_delete_custom_job(request_type=dict)
 
 
+def test_delete_custom_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.delete_custom_job), "__call__"
+    ) as call:
+        client.delete_custom_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.DeleteCustomJobRequest()
+
+
 @pytest.mark.asyncio
 async def test_delete_custom_job_async(
     transport: str = "grpc_asyncio", request_type=job_service.DeleteCustomJobRequest
@@ -1462,6 +1538,24 @@ def test_cancel_custom_job(
 
 def test_cancel_custom_job_from_dict():
     test_cancel_custom_job(request_type=dict)
+
+
+def test_cancel_custom_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.cancel_custom_job), "__call__"
+    ) as call:
+        client.cancel_custom_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.CancelCustomJobRequest()
 
 
 @pytest.mark.asyncio
@@ -1683,6 +1777,24 @@ def test_create_data_labeling_job(
 
 def test_create_data_labeling_job_from_dict():
     test_create_data_labeling_job(request_type=dict)
+
+
+def test_create_data_labeling_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.create_data_labeling_job), "__call__"
+    ) as call:
+        client.create_data_labeling_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.CreateDataLabelingJobRequest()
 
 
 @pytest.mark.asyncio
@@ -1959,6 +2071,24 @@ def test_get_data_labeling_job_from_dict():
     test_get_data_labeling_job(request_type=dict)
 
 
+def test_get_data_labeling_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.get_data_labeling_job), "__call__"
+    ) as call:
+        client.get_data_labeling_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.GetDataLabelingJobRequest()
+
+
 @pytest.mark.asyncio
 async def test_get_data_labeling_job_async(
     transport: str = "grpc_asyncio", request_type=job_service.GetDataLabelingJobRequest
@@ -2188,6 +2318,24 @@ def test_list_data_labeling_jobs(
 
 def test_list_data_labeling_jobs_from_dict():
     test_list_data_labeling_jobs(request_type=dict)
+
+
+def test_list_data_labeling_jobs_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.list_data_labeling_jobs), "__call__"
+    ) as call:
+        client.list_data_labeling_jobs()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.ListDataLabelingJobsRequest()
 
 
 @pytest.mark.asyncio
@@ -2563,6 +2711,24 @@ def test_delete_data_labeling_job_from_dict():
     test_delete_data_labeling_job(request_type=dict)
 
 
+def test_delete_data_labeling_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.delete_data_labeling_job), "__call__"
+    ) as call:
+        client.delete_data_labeling_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.DeleteDataLabelingJobRequest()
+
+
 @pytest.mark.asyncio
 async def test_delete_data_labeling_job_async(
     transport: str = "grpc_asyncio",
@@ -2760,6 +2926,24 @@ def test_cancel_data_labeling_job(
 
 def test_cancel_data_labeling_job_from_dict():
     test_cancel_data_labeling_job(request_type=dict)
+
+
+def test_cancel_data_labeling_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.cancel_data_labeling_job), "__call__"
+    ) as call:
+        client.cancel_data_labeling_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.CancelDataLabelingJobRequest()
 
 
 @pytest.mark.asyncio
@@ -2974,6 +3158,24 @@ def test_create_hyperparameter_tuning_job(
 
 def test_create_hyperparameter_tuning_job_from_dict():
     test_create_hyperparameter_tuning_job(request_type=dict)
+
+
+def test_create_hyperparameter_tuning_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.create_hyperparameter_tuning_job), "__call__"
+    ) as call:
+        client.create_hyperparameter_tuning_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.CreateHyperparameterTuningJobRequest()
 
 
 @pytest.mark.asyncio
@@ -3244,6 +3446,24 @@ def test_get_hyperparameter_tuning_job_from_dict():
     test_get_hyperparameter_tuning_job(request_type=dict)
 
 
+def test_get_hyperparameter_tuning_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.get_hyperparameter_tuning_job), "__call__"
+    ) as call:
+        client.get_hyperparameter_tuning_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.GetHyperparameterTuningJobRequest()
+
+
 @pytest.mark.asyncio
 async def test_get_hyperparameter_tuning_job_async(
     transport: str = "grpc_asyncio",
@@ -3466,6 +3686,24 @@ def test_list_hyperparameter_tuning_jobs(
 
 def test_list_hyperparameter_tuning_jobs_from_dict():
     test_list_hyperparameter_tuning_jobs(request_type=dict)
+
+
+def test_list_hyperparameter_tuning_jobs_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.list_hyperparameter_tuning_jobs), "__call__"
+    ) as call:
+        client.list_hyperparameter_tuning_jobs()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.ListHyperparameterTuningJobsRequest()
 
 
 @pytest.mark.asyncio
@@ -3858,6 +4096,24 @@ def test_delete_hyperparameter_tuning_job_from_dict():
     test_delete_hyperparameter_tuning_job(request_type=dict)
 
 
+def test_delete_hyperparameter_tuning_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.delete_hyperparameter_tuning_job), "__call__"
+    ) as call:
+        client.delete_hyperparameter_tuning_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.DeleteHyperparameterTuningJobRequest()
+
+
 @pytest.mark.asyncio
 async def test_delete_hyperparameter_tuning_job_async(
     transport: str = "grpc_asyncio",
@@ -4056,6 +4312,24 @@ def test_cancel_hyperparameter_tuning_job(
 
 def test_cancel_hyperparameter_tuning_job_from_dict():
     test_cancel_hyperparameter_tuning_job(request_type=dict)
+
+
+def test_cancel_hyperparameter_tuning_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.cancel_hyperparameter_tuning_job), "__call__"
+    ) as call:
+        client.cancel_hyperparameter_tuning_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.CancelHyperparameterTuningJobRequest()
 
 
 @pytest.mark.asyncio
@@ -4266,6 +4540,24 @@ def test_create_batch_prediction_job(
 
 def test_create_batch_prediction_job_from_dict():
     test_create_batch_prediction_job(request_type=dict)
+
+
+def test_create_batch_prediction_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.create_batch_prediction_job), "__call__"
+    ) as call:
+        client.create_batch_prediction_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.CreateBatchPredictionJobRequest()
 
 
 @pytest.mark.asyncio
@@ -4530,6 +4822,24 @@ def test_get_batch_prediction_job_from_dict():
     test_get_batch_prediction_job(request_type=dict)
 
 
+def test_get_batch_prediction_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.get_batch_prediction_job), "__call__"
+    ) as call:
+        client.get_batch_prediction_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.GetBatchPredictionJobRequest()
+
+
 @pytest.mark.asyncio
 async def test_get_batch_prediction_job_async(
     transport: str = "grpc_asyncio",
@@ -4748,6 +5058,24 @@ def test_list_batch_prediction_jobs(
 
 def test_list_batch_prediction_jobs_from_dict():
     test_list_batch_prediction_jobs(request_type=dict)
+
+
+def test_list_batch_prediction_jobs_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.list_batch_prediction_jobs), "__call__"
+    ) as call:
+        client.list_batch_prediction_jobs()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.ListBatchPredictionJobsRequest()
 
 
 @pytest.mark.asyncio
@@ -5127,6 +5455,24 @@ def test_delete_batch_prediction_job_from_dict():
     test_delete_batch_prediction_job(request_type=dict)
 
 
+def test_delete_batch_prediction_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.delete_batch_prediction_job), "__call__"
+    ) as call:
+        client.delete_batch_prediction_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.DeleteBatchPredictionJobRequest()
+
+
 @pytest.mark.asyncio
 async def test_delete_batch_prediction_job_async(
     transport: str = "grpc_asyncio",
@@ -5324,6 +5670,24 @@ def test_cancel_batch_prediction_job(
 
 def test_cancel_batch_prediction_job_from_dict():
     test_cancel_batch_prediction_job(request_type=dict)
+
+
+def test_cancel_batch_prediction_job_empty_call():
+    # This test is a coverage failsafe to make sure that totally empty calls,
+    # i.e. request == None and no flattened fields passed, work.
+    client = JobServiceClient(
+        credentials=credentials.AnonymousCredentials(), transport="grpc",
+    )
+
+    # Mock the actual call within the gRPC stub, and fake the request.
+    with mock.patch.object(
+        type(client.transport.cancel_batch_prediction_job), "__call__"
+    ) as call:
+        client.cancel_batch_prediction_job()
+        call.assert_called()
+        _, args, _ = call.mock_calls[0]
+
+        assert args[0] == job_service.CancelBatchPredictionJobRequest()
 
 
 @pytest.mark.asyncio